extract new system into modules for both commands
-
